apps/api/pyproject.toml
[project]
name = "tars-api"
version = "0.1.0"
description = "TARS - API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.16.5",
    "celery[redis,sqlalchemy]>=5.5.3",
    "fastapi[standard]>=0.118.2",
    "langchain>=1.0.0",
    "langchain-deepseek>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.0",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.5",
    "pymupdf4llm>=0.0.27",
    "qdrant-client>=1.15.1",
    "scapy>=2.6.1",
    "sqlmodel>=0.0.27",
    "tiktoken>=0.12.0",
    "typed-settings>=25.1.0",
]

[dependency-groups]
dev = ["flower>=2.0.1", "ruff>=0.14.0"]


[tool.alembic]

script_location = "%(here)s/alembic"


prepend_sys_path = ["."]

[[tool.alembic.post_write_hooks]]
# lint with attempts to fix using "ruff" - use the module runner, against the "ruff" module
name = "ruff"
type = "module"
module = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
